Create dataset from generator object (for decoding modes).

+++ g2p_seq2seq/app.py
@@ -31,7 +31,6 @@ import tensorflow as tf
 from g2p_seq2seq.g2p import G2PModel
 from g2p_seq2seq.params import Params
 
-from IPython.core.debugger import Tracer
 
 tf.flags.DEFINE_string("model_dir", None, "Training directory.")
 tf.flags.DEFINE_boolean("interactive", False,
@@ -65,8 +64,9 @@ tf.flags.DEFINE_integer("max_epochs", 0,
                         " (0: no limit).")
 tf.flags.DEFINE_integer("eval_steps", 10, "Number of steps for evaluation.")
 tf.flags.DEFINE_string("schedule", "train_and_evaluate",
-"""Set schedule. More info about training configurations you can read in
-tensor2tensor docs: https://github.com/tensorflow/tensor2tensor/blob/master/docs/distributed_training.md""")
+    """Set schedule. More info about training configurations you can read in
+    tensor2tensor docs: https://github.com/tensorflow/tensor2tensor/blob/master/
+docs/distributed_training.md""")
 tf.flags.DEFINE_string("master", "",
                        "TensorFlow master. Defaults to empty string for local."
                        "Specifies the configurations for distributed run.")
@@ -80,27 +80,34 @@ def main(_=[]):
   tf.logging.set_verbosity(tf.logging.INFO)
   data_path = FLAGS.train if FLAGS.train else FLAGS.decode
   data_path = FLAGS.evaluate if not data_path else data_path
+
+  if not FLAGS.model_dir:
+    raise RuntimeError("Model directory not specified.")
+
   if FLAGS.reinit and os.path.exists(FLAGS.model_dir):
     shutil.rmtree(FLAGS.model_dir)
 
+  if not os.path.exists(FLAGS.model_dir):
+    os.makedirs(FLAGS.model_dir)
+
   params = Params(FLAGS.model_dir, data_path, flags=FLAGS)
-  g2p_model = G2PModel(params)
 
   if FLAGS.train:
+    g2p_model = G2PModel(params, file_path=FLAGS.train, is_training=True)
     g2p_model.prepare_data(train_path=FLAGS.train, dev_path=FLAGS.valid)
     g2p_model.train()
 
   elif FLAGS.interactive:
+    g2p_model = G2PModel(params)
     g2p_model.interactive()
 
   elif FLAGS.decode:
-    g2p_model.prepare_data(test_path=FLAGS.decode)
-    g2p_model.decode(decode_file_path=FLAGS.decode,
-                     output_file_path=FLAGS.output)
+    g2p_model = G2PModel(params, file_path=FLAGS.decode, is_training=False)
+    g2p_model.decode(output_file_path=FLAGS.output)
 
   elif FLAGS.evaluate:
-    g2p_model.prepare_data(test_path=FLAGS.test)
-    g2p_model.evaluate(FLAGS.evaluate, FLAGS.test)
+    g2p_model = G2PModel(params, file_path=FLAGS.evaluate, is_training=False)
+    g2p_model.evaluate()
 
 
 if __name__ == "__main__":
